Add files via upload
This is a C++ implementation of a program that computes the TE and TM solutions for a slab waveguide. The program computes the polarisation dependent propagation constants and optical modes for a given waveguide. Also included is a python interface that can be used to call the slab_wg executable and make plots of the computed solutions.